requirements.txt
adlfs==2024.12.0
aiobotocore==2.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.21.0
azure-storage-blob==12.24.1
azure-storage-file-datalake==12.18.1
babel==2.17.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==1.2.0
boto3==1.35.81
botocore==1.35.93
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
dask==2024.12.1
dask-gateway==2024.1.0
databricks-sdk==0.41.0
databricks-sql-connector==4.0.0
datasketch==1.6.5
db-dtypes==1.4.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
delta-kernel-rust-sharing-wrapper==0.2.2
delta-sharing==1.3.1
dill==0.3.9
distributed==2024.12.1
et_xmlfile==2.0.0
executing==2.2.0
exrex==0.12.0
Faker==33.3.1
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fs-gcsfs==1.5.1
fsspec==2024.12.0
fuzzywuzzy==0.18.0
gcsfs==2024.12.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.31.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hdbscan==0.8.40
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
libsass==0.23.0
lightgbm==4.6.0
llvmlite==0.44.0
locket==1.0.0
lz4==4.4.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
miceforest==6.0.3
mistune==3.1.3
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.2.0
multimethod==1.12
murmurhash==1.0.12
mysql-connector-python==9.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.26.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pdf_reports==0.3.7
pexpect==4.9.0
phik==0.12.4
phonenumbers==8.13.55
pillow==11.1.0
platformdirs==4.3.7
preshed==3.0.9
presidio_analyzer==2.2.358
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyod==2.0.4
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyparsing==3.2.2
pyphen==0.17.2
pypugjs==5.11.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3fs==2024.12.0
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==78.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowflake-connector-python==3.14.0
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
tblib==3.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
thrift==0.20.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.3
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.64.1
traitlets==5.14.3
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
uuid==1.30
visions==0.7.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
yarl==1.18.3
ydata-core==0.7.0
ydata-datascience==0.7.0
ydata-profiling==4.16.1
ydata-sdk==2.1.0
zict==3.0.0
zopfli==0.2.3.post1
